Successfully add router and route to three separate views
