+++ src/components/AboutView.vue
@@ -1,0 +1,5 @@
+
+<template>
+
+    <p>About</p>
+</template>
