+++ src/components/Heading.vue
@@ -17,4 +17,3 @@ defineProps({
     </h3>
   </div>
 </template>
-
