+++ src/components/HomeView.vue
@@ -1,0 +1,8 @@
+<script>
+import Heading from "./Heading.vue";
+
+</script>
+<template>
+    <Heading />
+    <p>Home</p>
+</template>
