+++ src/components/Nav.vue
@@ -1,11 +1,14 @@
 
 
 <script>
-
+import { RouterLink } from 'vue-router';
 </script>
 
 <template>
     <nav>
+        <RouterLink to="/">Home</RouterLink>
+        <RouterLink to="about">About</RouterLink>
+        <RouterLink to="CSSColors">CSSColors</RouterLink>
         <a href="home">Home</a>
         <a href="about">About</a>
         <a href="CSSColors">CSSColors</a>
