+++ src/components/__tests__/HelloWorld.spec.js
@@ -1,7 +1,7 @@
 import { describe, it, expect } from 'vitest'
 
 import { mount } from '@vue/test-utils'
-import HelloWorld from '../HelloWorld.vue'
+import HelloWorld from '../Heading.vue'
 
 describe('HelloWorld', () => {
   it('renders properly', () => {
