+++ src/main.js
@@ -1,6 +1,9 @@
-import './assets/main.css'
+import './assets/main.css';
 
-import { createApp } from 'vue'
-import App from './App.vue'
+import { createApp } from 'vue';
+import App from './App.vue';
+import router from "./router/index.js";
 
-createApp(App).mount('#app')
+createApp(App)
+    .use(router)
+    .mount('#app')
